requirements.txt
flask==2.3.3
flask-cors==4.0.0
speechrecognition==3.10.0
pyttsx3==2.90
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
# pyaudio==0.2.11  # Optional: Install with brew install portaudio first
